And I doubt this only matched at line starts.
svn path=/branches/KDE/3.5/kdelibs/; revision=802658

+++ kdesu/process.cpp
@@ -507,13 +507,14 @@ int PtyProcess::waitForChild()
         if (ret) 
         {
             QCString output = readAll(false);
+            bool lineStart = true;
             while (!output.isNull()) 
             {
                 if (!m_Exit.isEmpty())
                 {
                     // match exit string only at line starts
                     int pos = output.find(m_Exit);
-                    if ((pos >= 0) && ((pos == 0) || (output.at (pos - 1) == '\n')))
+                    if ((pos >= 0) && ((pos == 0 && lineStart) || (output.at (pos - 1) == '\n')))
                     {
                         kill(m_Pid, SIGTERM);
                     }
@@ -523,6 +524,7 @@ int PtyProcess::waitForChild()
                     fputs(output, stdout);
                     fflush(stdout);
                 }
+                lineStart = output.at( output.length() - 1 ) == '\n';
                 output = readAll(false);
             }
         }
